Adjust folders.
