+++ DNAStrand/DNAStrand/View/GeneView.swift
